+++ src/main/java/forestry/api/core/IProductProducer.java
@@ -1,0 +1,13 @@
+package forestry.api.core;
+
+import java.util.List;
+
+/**
+ * Denotes a species that can produce products.
+ */
+public interface IProductProducer {
+	/**
+	 * @return The list of possible items that can be produced by this bee.
+	 */
+	List<IProduct> getProducts();
+}