+++ src/main/java/forestry/apiculture/compat/ApicultureJeiPlugin.java
@@ -1,7 +1,15 @@
 package forestry.apiculture.compat;
 
+import java.util.ArrayList;
+import java.util.List;
+
 import net.minecraft.resources.ResourceLocation;
 
+import forestry.api.ForestryConstants;
+import forestry.api.IForestryApi;
+import forestry.api.core.IProductProducer;
+import forestry.api.core.ISpecialtyProducer;
+import forestry.api.genetics.ISpeciesType;
 import forestry.api.genetics.alleles.BeeChromosomes;
 import forestry.api.modules.ForestryModuleIds;
 import forestry.apiculture.features.ApicultureItems;
@@ -10,14 +18,43 @@ import forestry.core.utils.SpeciesUtil;
 
 import mezz.jei.api.IModPlugin;
 import mezz.jei.api.JeiPlugin;
+import mezz.jei.api.gui.drawable.IDrawable;
+import mezz.jei.api.helpers.IGuiHelper;
+import mezz.jei.api.registration.IRecipeCategoryRegistration;
 import mezz.jei.api.registration.IRecipeRegistration;
 import mezz.jei.api.registration.ISubtypeRegistration;
 
 @JeiPlugin
 public class ApicultureJeiPlugin implements IModPlugin {
+	static final ResourceLocation BACKGROUND = ForestryConstants.forestry("textures/gui/jei/recipes.png");
+	static final List<ProductsRecipeCategory> productsCategories = new ArrayList<>();
+	static final List<MutationsRecipeCategory> mutationsCategories = new ArrayList<>();
+
 	@Override
 	public ResourceLocation getPluginUid() {
 		return ForestryModuleIds.APICULTURE;
+	}
+
+	@Override
+	public void registerCategories(IRecipeCategoryRegistration registration) {
+		IGuiHelper helper = registration.getJeiHelpers().getGuiHelper();
+		IDrawable productsBackground = helper.createDrawable(BACKGROUND, 0, 61 + 30, 162, 61);
+		IDrawable mutationsBackground = helper.createDrawable(BACKGROUND, 0, 0 + 30, 162, 61);
+
+		for (ISpeciesType<?, ?> type : IForestryApi.INSTANCE.getGeneticManager().getSpeciesTypes()) {
+			IDrawable icon = helper.createDrawableItemStack(type.createDefaultStack());
+
+			// products
+			if (type.getDefaultSpecies() instanceof IProductProducer) {
+				ProductsRecipeCategory category = new ProductsRecipeCategory(type, productsBackground, icon);
+				productsCategories.add(category);
+				registration.addRecipeCategories(category);
+			}
+			// mutations
+			MutationsRecipeCategory category = new MutationsRecipeCategory(type, mutationsBackground, icon);
+			mutationsCategories.add(category);
+			registration.addRecipeCategories(category);
+		}
 	}
 
 	@Override
@@ -44,5 +81,21 @@ public class ApicultureJeiPlugin implements IModPlugin {
 				ApicultureItems.HABITAT_LOCATOR,
 				ApicultureItems.SCOOP
 		);
+
+		for (MutationsRecipeCategory category : mutationsCategories) {
+			registry.addRecipes(category.getRecipeType(), category.speciesType.getMutations().getAllMutations().stream().map(MutationRecipe::new).toList());
+		}
+		for (ProductsRecipeCategory category : productsCategories) {
+			registry.addRecipes(category.getRecipeType(), category.speciesType.getAllSpecies().stream()
+					.filter(species -> {
+						// filter out species with no products or specialties
+						return (species instanceof IProductProducer producer && !producer.getProducts().isEmpty())
+								|| (species instanceof ISpecialtyProducer specialtyProducer && !specialtyProducer.getSpecialties().isEmpty());
+					})
+					.map(species -> new ProductRecipe(species.cast()))
+					.toList());
+		}
+		mutationsCategories.clear();
+		productsCategories.clear();
 	}
 }